requirements.txt
timm
marshmallow_dataclass
pyaml
click
Faker
hydra-core
pandas
opencv-python
scikit-learn
tensorboard
albumentations
mmcv